[Refactor] (test) Move the test spec class back to '.application'.
